bot/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
alembic==1.14.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
discord.py==2.4.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
Werkzeug==3.1.3
yarl==1.18.0
